requirements.txt
setuptools
dill
pyyaml
pytest
ensure==1.0.2
python-box==6.0.2
pathlib
json
zipfile